refactor(coral): rename API client module
